requirements.txt
python-dotenv
huggingface_hub
tavily-python
pydantic
streamlit
